cut over basic test
